Deleted AirportGUI Class Added Main Class
Decided to call it Main again and move methods into a controller package so as not to clutter this class and be more organized.
